Fixed "Mini Zoom".
